+++ lib/cocoapods/installer/xcode/pods_project_generator.rb
@@ -86,6 +86,11 @@ module Pod
           development_pod_targets.select(&:should_build?).each do |pod_target|
             next unless share_scheme_for_development_pod?(pod_target.pod_name)
             Xcodeproj::XCScheme.share_scheme(project.path, pod_target.label)
+            if pod_target.contains_test_specifications?
+              pod_target.supported_test_types.each do |test_type|
+                Xcodeproj::XCScheme.share_scheme(project.path, pod_target.test_target_label(test_type))
+              end
+            end
           end
         end
 
@@ -203,16 +208,10 @@ module Pod
               aggregate_target.native_target.add_dependency(pod_target.native_target)
               configure_app_extension_api_only_for_target(pod_target) if is_app_extension
 
-              pod_target.dependent_targets.each do |pod_dependency_target|
-                next unless pod_dependency_target.should_build?
-                pod_target.native_target.add_dependency(pod_dependency_target.native_target)
-                configure_app_extension_api_only_for_target(pod_dependency_target) if is_app_extension
+              add_dependent_targets_to_native_target(pod_target.dependent_targets, pod_target.native_target, is_app_extension, pod_target.requires_frameworks?, frameworks_group)
 
-                if pod_target.requires_frameworks?
-                  product_ref = frameworks_group.files.find { |f| f.path == pod_dependency_target.product_name } ||
-                    frameworks_group.new_product_ref_for_target(pod_dependency_target.product_basename, pod_dependency_target.product_type)
-                  pod_target.native_target.frameworks_build_phase.add_file_reference(product_ref, true)
-                end
+              pod_target.test_native_targets.each do |test_native_target|
+                add_dependent_targets_to_native_target(pod_target.test_dependent_targets, test_native_target, false, pod_target.requires_frameworks?, frameworks_group)
               end
             end
           end
@@ -250,6 +249,20 @@ module Pod
         # @! group Private Helpers
 
         private
+
+        def add_dependent_targets_to_native_target(dependent_targets, native_target, is_app_extension, requires_frameworks, frameworks_group)
+          dependent_targets.each do |pod_dependency_target|
+            next unless pod_dependency_target.should_build?
+            native_target.add_dependency(pod_dependency_target.native_target)
+            configure_app_extension_api_only_for_target(pod_dependency_target) if is_app_extension
+
+            if requires_frameworks
+              product_ref = frameworks_group.files.find { |f| f.path == pod_dependency_target.product_name } ||
+                  frameworks_group.new_product_ref_for_target(pod_dependency_target.product_basename, pod_dependency_target.product_type)
+              native_target.frameworks_build_phase.add_file_reference(product_ref, true)
+            end
+          end
+        end
 
         # Sets the APPLICATION_EXTENSION_API_ONLY build setting to YES for all
         # configurations of the given target