+++ lib/cocoapods/resolver.rb
@@ -305,7 +305,7 @@ module Pod
       find_cached_set(dependency).
         all_specifications(installation_options.warn_for_multiple_pod_sources).
         select { |s| requirement.satisfied_by? s.version }.
-        map { |s| s.subspec_by_name(dependency.name, false) }.
+        map { |s| s.subspec_by_name(dependency.name, false, true) }.
         compact.
         reverse
     end